src/sboannotator/requirements.txt
setuptools>=58.0.4
cobra>=0.22.1
python-libsbml>=5.19.0
pandas>=2.0.3
requests>=2.28.2